requirements.txt
streamlit
folium
streamlit-folium
pandas
matplotlib
fpdf
supabase
requests
Pillow